[refactor] NumberGenerator 클래스 생성 및 불필요 클래스 제거
